Add soft-failing Complement to the Synapse pipeline (#97)
This PR adds a Complement build step to the Synapse pipeline. It is currently marked as "soft-fail" (meaning this step failing will not block the whole build) as not all complement tests work with Synapse yet.

It is modeled after the existing Dendrite pipeline entry, but instead moves the commands to a script inside the Synapse repo.